// frontend/src/codegen/model/permissionStatsSnapshotResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Flathub API
 * OpenAPI spec version: 0.1.0
 */
import type { PermissionStatsSnapshotResultPermissionCounts } from "./permissionStatsSnapshotResultPermissionCounts"

export interface PermissionStatsSnapshotResult {
  snapshot_date: string
  eligible_apps: number
  apps_with_stable_metadata: number
  permission_counts: PermissionStatsSnapshotResultPermissionCounts
}