// frontend/src/codegen/model/permissionStatsWindowResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Flathub API
 * OpenAPI spec version: 0.1.0
 */
import type { PermissionStatsSnapshotResult } from "./permissionStatsSnapshotResult"

export interface PermissionStatsWindowResult {
  start_month: string
  end_month: string
  snapshots: PermissionStatsSnapshotResult[]
}